requirements.txt
python-telegram-bot[job-queue]==20.7
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.7
telethon==1.34.0
python-docx==1.2.0
openpyxl==3.1.2